requirements.txt
langchain
langchain-community
langchain-openai
openai
chromadb
PyPDF2
fastapi
uvicorn
python-dotenv
pypdf
tiktoken